Relational mapping changes
eliminating unnecessary load for comment fetch within the post data

+++ SMP-server/src/main/java/com/smp/main/service/CommentService.java
@@ -1,7 +1,8 @@
 package com.smp.main.service;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList; 
+import java.util.ArrayList;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,9 +12,11 @@ import com.smp.main.exceptions.PostNotExists;
 import com.smp.main.exceptions.UserNotExists;
 import com.smp.main.model.Comment;
 import com.smp.main.model.Post;
+import com.smp.main.model.Reel;
 import com.smp.main.model.User;
 import com.smp.main.repository.CommentRepository;
 import com.smp.main.repository.PostRepository;
+import com.smp.main.repository.ReelRepository;
 import com.smp.main.repository.UserRepository;
 @Service
 public class CommentService {
@@ -25,14 +28,15 @@ public class CommentService {
 	UserRepository userRepository;
 	@Autowired 
 	PostRepository postRepository;
-
-	public Comment createComment(String content,Long userId,String postId) {
+	@Autowired 
+	ReelRepository reelRepository;
+	public Comment createPostComment(String content,Long userId,String postId) {
 	
 		User user= userRepository.findById(userId).orElse(null);
 		Post post=postRepository.findById(postId).orElse(null);
 		if(user!=null && post!=null) {
-			Comment comment=commentRepository.save(new Comment(content,new ArrayList<>(),userId,postId,LocalDateTime.now()));
-			post.getComments().add(comment);
+			Comment comment=commentRepository.save(new Comment(content,new ArrayList<>(),userId,postId,user.getName(),LocalDateTime.now()));
+			post.getComments().add(comment.getId());
 			postRepository.save(post);
 			return comment;
 		}
@@ -42,6 +46,25 @@ public class CommentService {
 			else
 				throw new PostNotExists("Post not found");
 		}
+	}
+	
+
+	public Comment createReelComment(String content, Long userId, String reelId) {
+
+		User user= userRepository.findById(userId).orElse(null);
+		Reel reel=reelRepository.findById(reelId).orElse(null);
+		if(user!=null && reel!=null) {
+			Comment comment=commentRepository.save(new Comment(content,new ArrayList<>(),userId,reelId,user.getName(),LocalDateTime.now()));
+			reel.getComments().add(comment.getId());
+			reelRepository.save(reel);
+			return comment;
+		}
+		else {
+			if(user==null)
+			throw new UserNotExists("User not found.");
+			else
+				throw new PostNotExists("Post not found");
+		} 
 	}
 	//comment liked by users
 	public Comment likeComment(String commentID,Long userID) {
@@ -74,16 +97,25 @@ public class CommentService {
 		
 	}
 	//delete comment
-	public String deleteComment(String commentID) {
+	public String deleteComment(String commentID, User user) {
 		Comment comment=commentRepository.findById(commentID).orElse(null);
-		if(comment!=null) {
+		if(comment!=null && user!=null && user.getId()==comment.getUserId()) {
 			commentRepository.deleteById(commentID);
 			return "Successfully deleted";
 		}
 		else {
-			throw new CommentNotExists("Comment not found");
+			throw new CommentNotExists("Comment not found with this user");
 		}
 		
+	}
+	public List<Comment> getCommentByPostId(String postId) {
+		List<Comment> comments= commentRepository.findByPostId(postId);
+		return comments;
+	}
+
+
+	public List<Comment> getComments(List<String> cId) { 
+		return commentRepository.findAllById(cId);
 	}
 
 }
